Delete separate line for answer
The calculation will be performed on a single line in one place. That is, if the user inputs '2+2' and clicks '=', the display is going to show, consequtively, just the '2', then another '2' and then the 4. This is instead the previous idea for implementation, whereby it would print the entirety of '2+2' and display '4' on a separate line.