Fix output when no operator is present
The programme would glitch if asked to compute (i.e., if '=' was pressed) but no operator was given. The glitch was due to the programme searching for the operator and second argument, which in this case were non-existent. The issue was fixed; now the output stays unchanged when the user presses '=' having not had entered an operator and second argument.